Refactor: Extract routes to separate files
This commit extracts the /mutate, /select, and /mutators routes into their own files to improve code organization and maintainability.
